refactor : 위경도 자료형 변경

+++ Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/common/vo/CoordinateVo.kt
@@ -4,6 +4,6 @@ import javax.persistence.Embeddable
 
 @Embeddable
 class CoordinateVo(
-    var latitude: Long, // 위도
-    var longitude: Long, // 경도
+    var latitude: Double, // 위도
+    var longitude: Double, // 경도
 )
